Move class out to seperate file
