+++ java/statestore/src/main/java/sleeper/statestore/transactionlog/snapshots/TransactionLogSnapshotDeleter.java
@@ -105,42 +105,4 @@ public class TransactionLogSnapshotDeleter {
          */
         void delete(String path) throws IOException;
     }
-
-    /**
-     * ADD Stuff here.
-     */
-    public class DeleteProcesState {
-        Integer deletedCount = 0;
-        Long lastTransactionNumber = Long.MIN_VALUE;
-
-        /**
-         * Store key details on the success of a deletion process. Increments the count of the deletion and stores
-         * the transaction number for the last instacen for use in logging.
-         *
-         * @param transactionNumber most recent transaction number deleted, done so as to save the last one for logging
-         */
-        public void deleteSuccess(long transactionNumber) {
-            this.lastTransactionNumber = transactionNumber;
-            deletedCount++;
-        }
-
-        public Integer getDeletedCount() {
-            return deletedCount;
-        }
-
-        /**
-         * Method for returning the number of the last transaction deleted, is overwritten on ever call as only
-         * latest value needed.
-         *
-         * @return if a trnasaction has been deleted then the id of the transaction, if none have it returns null
-         */
-        public Long getLastTransactionNumber() {
-            if (lastTransactionNumber != Long.MIN_VALUE) {
-                return lastTransactionNumber;
-            } else {
-                return null;
-            }
-        }
-
-    }
 }
